requirements.txt
pyserial
paho-mqtt
Pillow
python-dotenv
pydantic
pydantic-settings
pygame
setuptools
numpy